\docType{data}
\name{allele.freqs}
\alias{allele.freqs}
\title{Sample allele frequencies}
\format{Tabular format. The records represent unique SSR alleles with their assigned frequencies.
Frequencies were derived from the sample data and are just for illustrative purposes.
\itemize{
 \item{"marker"} {Marker name}
 \item{"marker_size"} {Marker size}
 \item{"frequency"} {A fraction between 0 and 1.}
}}
\description{
Sample allele frequencies
}
\keyword{datasets}

