\docType{data}
\name{potato.quipu}
\alias{potato.quipu}
\title{SSR sample data for a set of potato accessions}
\format{Tabular format. The records represent unique SSR marker weights in base pairs as obtained
   for a set of three accessions. The combination of the first three columns is unique. The fourth
   column map_location is used for assigning markers to chromosomes or linkage groups.
\itemize{
 \item{"acccession_id"} {Accession ID}
 \item{"marker"} {Marker name}
 \item{"marker_size"} {Marker size}
 \item{"map_location"} {Genetic ap location; usually Roman numbers for chromosomes or linkage group.}
}}
\description{
SSR sample data for a set of potato accessions
}
\keyword{datasets}

