\docType{package}
\name{quipu-package}
\alias{quipu-package}
\title{Quipu-type charts for a set of SSR markers.}
\description{
The chart shows SSR marker weights on a linear scale where
each allele or 'gel band' is represented by a circle. The
circle's diameter is sized inversely by its rareness within
the set of accessions in the database at hand.
}
\details{
Motivation: Genebanks increasingly use molecular markers
for routine characterization of ex-situ collections and
farmer managed diversity. CIP's (International Potato
Center) genebank presently uses a SSR marker-kit to produce
molecular profiles for potato accessions. We have been
searching for a compact graphical representation that shows
both molecular diversity and accession characteristics -
thus permitting biologists and collection curators to have
a simpler way to interpret high-volume data. Inspired by
the ancient Andean quipus we devised a graph that allows
for standardized representation while leaving room for
updates of the marker kit and the collection of accessions.
The graph has been used in several CIP publications.
}

