% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quipu.R
\name{runDemo}
\alias{runDemo}
\title{Run a short interactive demo}
\usage{
runDemo()
}
\description{
Shows the two typical plots and the effects of the main parameters.
}
\examples{
library(quipu)

if(interactive()) {
  runDemo()
}
}
\author{
Reinhard Simon
}
